Add files via upload
Minor changes to main RSBot, tweaks for consistency mainly. Created "CurserScan" which contains a method that drags the cursor around the screen, intended to be used as a tool to search the screen with the mouse. Also contains a variant of the screenshot method, but this one takes a picture of a select area on the screen. Plan to tweak this later to accept the rectangle position and dimensions for easier use