Add files via upload
Improved portal scan accuracy, removed constructors. created methods to scan the inventory as well as determine if it is full of rune essence. Initially had the FullEssenceCheck return a boolean, the turned it into a loop that breaks when a full inventory is detected. Will most likely add constructor to determine how long the inventory will be scanned before breaking regardless.